Refactor edit team page to use new layout
* Refactor edit team members page to use new layout

* Move edit team page into modules

* Remove legacy code
