--- src/apps/investments/views/team/edit-team-members.njk
@@ -1,8 +1,0 @@
-{% extends "../template.njk" %}
-
-{% block main_grid_right_column %}
-    {% component 'react-slot', {
-        id: 'edit-team-members',
-        props: props
-    } %}
-{% endblock %}
