+++ src/client/modules/Investments/Projects/tasks.js
@@ -5,6 +5,7 @@ import {
   transformResponseToCompanyCollection,
 } from './transformers'
 import { UNITED_KINGDOM_ID } from '../../../../common/constants'
+import { validateTeamAdvisersAreUnique } from './validator'
 
 export const updateInvestmentProject = (values) =>
   apiProxyAxios.patch(`v3/investment/${values.id}`, values)
@@ -78,3 +79,17 @@ export const updateRecipientCompany = (values) =>
   apiProxyAxios.patch(`v3/investment/${values.projectId}`, {
     uk_company: values.companyId ? values.companyId : null,
   })
+
+export const updateTeamMembers = ({ teamMembers, id }) => {
+  return new Promise(async (resolve, reject) => {
+    const reasonAdvisersAreNotUnique =
+      validateTeamAdvisersAreUnique(teamMembers)
+    if (reasonAdvisersAreNotUnique) {
+      reject(reasonAdvisersAreNotUnique)
+    } else {
+      const url = `v3/investment/${id}/team-member`
+      const response = await apiProxyAxios.put(url, teamMembers)
+      resolve(response)
+    }
+  })
+}
