+++ src/client/routes.js
@@ -40,6 +40,7 @@ import ProjectInteractions from './modules/Investments/Projects/ProjectInteracti
 import ProjectPropositions from './modules/Investments/Projects/ProjectPropositions'
 import EditClientRelationshipManagement from './modules/Investments/Projects/EditClientRelationshipManagement'
 import EditProjectManagement from './modules/Investments/Projects/EditProjectManagement'
+import EditTeamMembers from './modules/Investments/Projects/EditTeamMembers'
 
 const routes = {
   companies: [
@@ -286,6 +287,11 @@ const routes = {
       path: '/investments/projects/:projectId/edit-project-management',
       module: 'datahub:investments',
       component: EditProjectManagement,
+    },
+    {
+      path: '/investments/projects/:projectId/edit-team-members',
+      module: 'datahub:investments',
+      component: EditTeamMembers,
     },
   ],
 }
